fix: use PtrInt for pointer params in engine exports on 64-bit native target
i32 pointer params truncate 64-bit stack addresses. PtrInt resolves to
i32 on WASM and usize on native so the same export signatures work on
both targets without casting. gengo-engine.h updated to use uintptr_t.

+++ gengo-engine.h
@@ -35,9 +35,9 @@ typedef struct {
 /* ── Host module function descriptor ──────────────────────────────────────── */
 
 typedef struct {
-    uint32_t name_ptr;
-    uint32_t name_len;
-    uint32_t arity;
+    uintptr_t name_ptr; /* pointer to the function name string */
+    uint32_t  name_len;
+    uint32_t  arity;
 } gengo_host_module_func_def_t;
 
 /* ── Write callback (native target only) ──────────────────────────────────── */
